+++ generators/app/templates/shared/assets/app.js
@@ -1,0 +1,16 @@
+/* ========================================
+ * Application
+ * ======================================== */
+
+import $ from 'jquery'
+import fastClick from 'fastclick'
+
+/**
+ * Components
+ */
+// import './components/elements/SomeElement'
+
+/**
+ * Scripts
+ */
+fastClick(document.body)