+++ generators/app/templates/wordpress/_gulpfile.js
@@ -6,24 +6,22 @@
  * @url <%= (generatorRepository) %>
  */
 
-var assemble = require('assemble')
-var app = assemble()
 var fs = require('fs')
 var browserify = require('browserify')
 var browserSync = require('browser-sync').create()
-var handlebarsHelpers = require('handlebars-helpers')()
+var gulp = require('gulp')
 var notifier = require('node-notifier')
 var path = require('path')
 var prettyHrtime = require('pretty-hrtime')
 var rimraf = require('rimraf')
 var source = require('vinyl-source-stream')
 var through = require('through2')
-var watch = require('base-watch')
 var watchify = require('watchify')
 var $ = require('gulp-load-plugins')()
 
 var production = process.env.NODE_ENV === 'production'
-var open = process.env.npm_config_open
+var host = process.env.npm_config_host
+var open = process.env.npm_config_disable_open ? false : 'external'
 
 
 /* ======
@@ -31,43 +29,33 @@ var open = process.env.npm_config_open
  * ====== */
 
 var config = {
-  src: 'src/',
+  theme: '/',
+  host: '<%= dir %>.dev:8000',
+  src: './',
   dest: 'dist/',
-  buildPath: '/',
   stylesheets: {
-    src: 'src/assets/stylesheets/app.scss',
-    dest: 'dist/assets/stylesheets/',
-    watch: 'src/assets/stylesheets/**/**/*.scss'
+    src: 'assets/app.scss',
+    dest: 'dist/assets/',
+    watch: 'assets/**/**/*.scss'
   },
   javascripts: {
-    src: 'src/assets/javascripts/',
-    dest: 'dist/assets/javascripts/',
+    src: 'assets/',
+    dest: 'dist/assets/',
     bundle: [{
-      file_name: 'app.js',
-      src: 'src/assets/javascripts/app.js'
+      fileName: 'app.js'
     }, {
-      file_name: 'head.js',
-      src: 'src/assets/javascripts/head.js'
+      fileName: 'app.head.js'
     }]
   },
   images: {
-    src: 'src/assets/images/*.{jpg,jpeg,png,gif,webp,svg}',
+    src: 'assets/images/*.{jpg,jpeg,png,gif,webp,svg}',
     dest: 'dist/assets/images/',
-    watch: 'src/assets/images/*.{jpg,jpeg,png,gif,webp,svg}'
+    watch: 'assets/images/*.{jpg,jpeg,png,gif,webp,svg}'
   },
   fonts: {
-    src: 'src/assets/fonts/*.{eot,svg,ttf,woff,woff2}',
+    src: 'assets/fonts/*.{eot,svg,ttf,woff,woff2}',
     dest: 'dist/assets/fonts/',
-    watch: 'src/assets/fonts/*.{eot,svg,ttf,woff,woff2}'
-  },
-  html: {
-    data: 'src/*.{json,yml}',
-    dest: 'dist/',
-    helpers: 'src/helpers/*.js',
-    layouts: 'src/layouts/*.hbs',
-    partials: 'src/partials/*.hbs',
-    templates: 'src/templates/**/*.hbs',
-    watch: ['src/{layouts,templates,partials}/**/*.hbs', 'src/*.{json,yml}']
+    watch: 'assets/fonts/*.{eot,svg,ttf,woff,woff2}'
   }
 }
 
@@ -77,36 +65,10 @@ var config = {
  * ====== */
 
 /**
- * Assemble
- */
-app.data({ assets: 'assets' })
-app.data(config.html.data)
-app.helpers(config.html.helpers)
-app.helpers(handlebarsHelpers)
-app.use(watch())
-
-app.preLayout(/\/src\/templates\/.*\.hbs$/, function(view, next) {
-  view.layout = 'default'
-  next()
-})
-
-app.task('html', function() {
-  app.data({ dev: !production })
-  app.layouts(config.html.layouts)
-  app.partials(config.html.partials)
-
-  return app.src(config.html.templates)
-    .pipe(app.renderFile())
-    .on('error', handleError)
-    .pipe($.rename({ extname: '.html' }))
-    .pipe(app.dest(config.html.dest))
-})
-
-/**
  * Stylesheets
  */
-app.task('stylesheets', function() {
-  var pipeline = app.src(config.stylesheets.src)
+gulp.task('stylesheets', function() {
+  var pipeline = gulp.src(config.stylesheets.src)
     .pipe($.sass({
       includePaths: ['node_modules', 'bower_components'],
       outputStyle: 'expanded'
@@ -119,13 +81,13 @@ app.task('stylesheets', function() {
 
   if (production) {
     return pipeline = pipeline
-      .pipe($.replace('../', config.buildPath + 'assets/'))
+      .pipe($.replace('./', '/wp-content/themes/' + config.theme + config.dest + 'assets/'))
       .pipe($.combineMq({ beautify: false }))
-      .pipe($.cssnano({ mergeRules: false }))
-      .pipe(app.dest(config.stylesheets.dest))
+      .pipe($.cssnano({ mergeRules: false, zindex: false }))
+      .pipe(gulp.dest(config.stylesheets.dest))
   } else {
     return pipeline = pipeline
-      .pipe(app.dest(config.stylesheets.dest))
+      .pipe(gulp.dest(config.stylesheets.dest))
       .pipe(browserSync.stream())
   }
 })
@@ -133,7 +95,7 @@ app.task('stylesheets', function() {
 /**
  * Javascripts
  */
-app.task('javascripts', function(callback) {
+gulp.task('javascripts', ['modernizr'], function(callback) {
 
   var bundleQueue = config.javascripts.bundle.length
 
@@ -143,17 +105,17 @@ app.task('javascripts', function(callback) {
       cache: {},
       packageCache: {},
       fullPaths: false,
-      entries: bundleConfig.src,
+      entries: config.javascripts.src + bundleConfig.fileName,
       debug: !production
     })
 
     var bundle = function() {
-      bundleLogger.start(bundleConfig.file_name)
+      bundleLogger.start(bundleConfig.fileName)
 
       var collect = pipeline
         .bundle()
         .on('error', handleError)
-        .pipe(source(bundleConfig.file_name))
+        .pipe(source(bundleConfig.fileName))
 
       if (!production) {
         collect = collect.pipe(browserSync.stream())
@@ -162,7 +124,7 @@ app.task('javascripts', function(callback) {
       }
 
       return collect
-        .pipe(app.dest(config.javascripts.dest))
+        .pipe(gulp.dest(config.javascripts.dest))
         .on('end', reportFinished)
     }
 
@@ -171,7 +133,7 @@ app.task('javascripts', function(callback) {
     }
 
     var reportFinished = function() {
-      bundleLogger.end(bundleConfig.file_name)
+      bundleLogger.end(bundleConfig.fileName)
 
       if (bundleQueue) {
         bundleQueue--
@@ -190,8 +152,8 @@ app.task('javascripts', function(callback) {
 /**
  * Images
  */
-app.task('images', function() {
-  return app.src(config.images.src)
+gulp.task('images', function() {
+  return gulp.src(config.images.src)
     .pipe($.changed(config.images.dest))
     .pipe($.imagemin({
       svgoPlugins: [
@@ -199,74 +161,63 @@ app.task('images', function() {
       ],
     }))
     .on('error', handleError)
-    .pipe(app.dest(config.images.dest))
+    .pipe(gulp.dest(config.images.dest))
 })
 
 /**
  * Fonts
  */
-app.task('fonts', function() {
-  return app.src(config.fonts.src)
+gulp.task('fonts', function() {
+  return gulp.src(config.fonts.src)
     .pipe($.changed(config.fonts.dest))
     .on('error', handleError)
-    .pipe(app.dest(config.fonts.dest))
+    .pipe(gulp.dest(config.fonts.dest))
 })
 
 /**
  * Server
  */
-app.task('server', function() {
+gulp.task('server', function() {
   browserSync.init({
-    open: open === undefined ? 'external' : open,
+    open: open,
     port: 9001,
+    proxy: host ? host : config.host,
     notify: false,
-    server: {
-      baseDir: config.dest,
-      routes: {
-        '/bower_components': 'bower_components',
-        '/node_modules': 'node_modules'
-      }
-    }
+    serveStatic: ['./']
   })
 })
 
 /**
  * Watch
  */
-app.task('watch', function() {
-  app.watch(config.html.watch, ['html'], function(cb) {
-    setTimeout(function() {
-      browserSync.reload()
-      cb()
-    }, 150)
-  })
-  app.watch(config.stylesheets.watch, ['stylesheets'])
-  app.watch(config.fonts.watch, ['fonts'])
-  app.watch(config.images.watch, ['images'])
+gulp.task('watch', function(callback) {
+  gulp.watch(config.src + '**/*.php').on('change', browserSync.reload)
+  gulp.watch(config.stylesheets.watch, ['stylesheets'])
+  gulp.watch(config.fonts.watch, ['fonts'])
+  gulp.watch(config.images.watch, ['images'])
 })
 
 /**
  * JavasScript Coding style
  */
-app.task('eslint', function () {
-  return app.src(config.javascripts.src + '**/*.js')
-    .pipe($.eslint())
-    .pipe($.eslint.format())
-    .pipe($.eslint.failAfterError())
-    .pipe(app.dest(config.javascripts.src))
-})
+ gulp.task('eslint', function () {
+   return gulp.src(config.javascripts.src + '**/*.js')
+     .pipe($.eslint())
+     .pipe($.eslint.format())
+     .pipe($.eslint.failAfterError())
+ })
 
 /**
  * Modernizr
  */
-app.task('modernizr', function() {
-  return app.src([
+gulp.task('modernizr', ['stylesheets'], function() {
+  return gulp.src([
     config.javascripts.src + '**/*.js',
     config.stylesheets.dest + 'app.css'
   ])
     .pipe($.modernizr({
       excludeTests: ['hidden'],
-      tests: [''],
+      tests: ['objectfit'],
       options: [
         'setClasses',
         'addTest',
@@ -277,57 +228,67 @@ app.task('modernizr', function() {
       ]
     }))
     .on('error', handleError)
-    .pipe(app.dest(config.javascripts.dest + 'vendors'))
+    .pipe(gulp.dest(config.javascripts.dest + 'vendors'))
 })
 
 /**
- * Inline <head> css/js
+ * Combined tasks
  */
-app.task('inline', function() {
-  return app.src([
-    config.dest + '**/*.html'
-  ], { base: config.dest })
+var tasks = ['stylesheets', 'javascripts', 'images', 'fonts']
+
+/**
+ * Create dist files and inline <head> css/js
+ */
+gulp.task('createDistPartials', tasks, function() {
+  return gulp.src([
+    config.src + 'partials/top.php',
+    config.src + 'partials/bottom.php',
+  ], { base: config.src })
     .pipe($.replace(inline({ matchFile: 'app.css' }), function() {
       return inline({ file: 'app.css' })
     }))
-    .pipe($.replace(inline({ matchFile: 'head.js' }), function() {
-      return inline({ file: 'head.js' })
+    .pipe($.replace(inline({ matchFile: 'app.head.js' }), function() {
+      return inline({ file: 'app.head.js' })
     }))
-    .pipe(app.dest(config.dest))
+    .pipe($.rename({ suffix: '.dist' }))
+    .pipe(gulp.dest(config.dest))
 })
 
 /**
  * Revision and remove unneeded files
  */
-app.task('rev', function() {
-  rimraf.sync(config.stylesheets.dest)
-  rimraf.sync(config.javascripts.dest + 'head.js')
-  rimraf.sync(config.javascripts.dest + 'vendors')
+gulp.task('rev', tasks.concat(['createDistPartials']), function() {
+  rimraf.sync(config.stylesheets.dest + '*.css')
+  rimraf.sync(config.javascripts.dest + 'app.head.js')
+  rimraf.sync(config.javascripts.dest + 'vendors/')
 
-  return app.src([
-    config.dest + 'assets/{javascripts,images,fonts}/**'
+  return gulp.src([
+    config.dest + 'assets/*.js',
+    config.dest + 'assets/{images,fonts}/**'
   ])
     .pipe($.rev())
-    .pipe(app.dest(config.dest + 'assets/'))
+    .pipe(gulp.dest(config.dest + 'assets'))
     .pipe(rmOriginalFiles())
     .pipe($.rev.manifest())
-    .pipe(app.dest('./'))
+    .pipe(gulp.dest('./'))
 })
 
 /**
  * Update references
  */
-app.task('updateReferences', function() {
-  var manifest = app.src('./rev-manifest.json')
+gulp.task('updateReferences', tasks.concat(['rev']), function() {
+  var manifest = gulp.src('./rev-manifest.json')
 
-  return app.src([
-    config.dest + '**'
+  return gulp.src([
+    config.dest + 'assets/**',
+    config.dest + 'partials/top.dist.php',
+    config.dest + 'partials/bottom.dist.php'
   ], { base: config.dest })
     .pipe($.revReplace({
       manifest: manifest,
-      replaceInExtensions: ['.js', '.css', '.html']
+      replaceInExtensions: ['.js', '.css', '.php']
     }))
-    .pipe(app.dest(config.dest))
+    .pipe(gulp.dest(config.dest))
 })
 
 
@@ -335,26 +296,24 @@ app.task('updateReferences', function() {
  * Main collected tasks
  * ====== */
 
-var tasks = ['stylesheets', 'modernizr', 'javascripts', 'images', 'fonts']
-
-app.task('build', ['eslint'], function() {
+gulp.task('build', ['eslint'], function() {
   rimraf.sync(config.dest)
-  app.build(tasks.concat([
-    'html',
-    'inline',
+  gulp.start(tasks.concat([
+    'modernizr',
+    'createDistPartials',
     'rev',
     'updateReferences'
-  ]), function(err) {
-    if (err) throw err
-  })
+  ]))
 })
 
-app.task('default', ['build'])
+gulp.task('default', ['build'])
 
-app.task('dev', function() {
-  rimraf.sync(config.dest)
-  app.build(tasks.concat(['html']), app.parallel(['server', 'watch']))
-})
+gulp.task('dev', tasks.concat([
+  'modernizr',
+  'watch',
+  'server'
+]))
+
 
 /* ======
  * Utilities
@@ -413,12 +372,10 @@ function rmOriginalFiles() {
   return through.obj(function(file, enc, cb) {
 
     if (file.revOrigPath) {
-      fs.unlink(file.revOrigPath)
+      fs.unlinkSync(file.revOrigPath)
     }
 
     this.push(file)
     return cb()
   })
 }
-
-module.exports = app