+++ generators/app/templates/wordpress/partials/bottom.php
@@ -1,0 +1,3 @@
+<script src="<?php bloginfo('template_directory'); ?>/dist/assets/app.js"></script>
+
+<?php wp_footer(); ?>