+++ generators/app/templates/wordpress/partials/top.php
@@ -1,0 +1,2 @@
+<script src="<?php bloginfo('template_directory'); ?>/dist/assets/app.head.js"></script>
+<link href="<?php bloginfo('template_directory'); ?>/dist/assets/app.css" rel="stylesheet" />